test: split over-cap routing fixtures
